add common test base
